+++ src/DataRepository/AbstractCephDataRepository.php
@@ -35,15 +35,4 @@ abstract class AbstractCephDataRepository extends BasicObjectDataRepository
             ->getClassMetadata($this->getClassName())
             ->getIdentifierFieldNames();
     }
-
-    protected function checkLimitAndOffset(?int $limit, ?int $offset): void
-    {
-        if ($limit !== null && $limit < 1) {
-            throw new \InvalidArgumentException(sprintf("Limit %d is not valid", $limit));
-        }
-
-        if ($offset && $offset < 0) {
-            throw new \InvalidArgumentException(sprintf("Offset %d is not valid", $offset));
-        }
-    }
 }